feat(library): change terms
