+++ .github/workflows/mc-release.yml
@@ -50,7 +50,7 @@ jobs:
           echo "---------------------------||---------------------------"
           echo "---------------------------||---------------------------"
           echo "---------------------------||---------------------------"
-          ./gradlew :mark_authenticator:publishMavenPublicationToMavenCentralRepository
+          ./gradlew :mark_biometric:publishMavenPublicationToMavenCentralRepository
       - name: Create Release From TAG
         uses: actions/create-release@v1
         env:
