+++ .pr_agent.toml
@@ -8,7 +8,7 @@ push_commands = [
 enable_chat_in_code_suggestions = true
 
 [config]
-allow_only_specific_folders=['.github','app', 'gradle', 'mark_authenticator']
+allow_only_specific_folders=['.github','app', 'gradle', 'mark_biometric']
 
 [review_agent]
 issues_user_guidelines = """
