--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/api/MarkAuthenticator.kt
@@ -1,264 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.api
-
-import android.content.Context
-import android.os.Build
-import androidx.annotation.RequiresApi
-import androidx.fragment.app.FragmentActivity
-import com.fadlurahmanfdev.mark_authenticator.api.callback.SecureAuthenticationDecryptCallback
-import com.fadlurahmanfdev.mark_authenticator.api.callback.SecureAuthenticationEncryptCallback
-import com.fadlurahmanfdev.mark_authenticator.api.callback.WeakAuthenticationCallback
-import com.fadlurahmanfdev.mark_authenticator.domain.enums.MarkAuthenticationStatus
-import com.fadlurahmanfdev.mark_authenticator.domain.enums.MarkAuthenticatorMethod
-import com.fadlurahmanfdev.mark_authenticator.internal.composition.MarkAuthenticatorCompositionRoot
-import javax.crypto.Cipher
-import javax.crypto.SecretKey
-
-/**
- * Public entry point for Mark Authenticator.
- *
- * This class delegates all operations to the internal implementation while keeping a stable,
- * consumer-friendly API.
- */
-class MarkAuthenticator(context: Context) : MarkAuthenticatorApi {
-    private val delegate: MarkAuthenticatorApi = MarkAuthenticatorCompositionRoot.create(context)
-
-    /**
-     * Generates a new Android KeyStore secret key for [alias].
-     *
-     * @param invalidatedByBiometricEnrollment if `true`, biometric enrollment changes invalidate key.
-     */
-    override fun generateSecretKey(
-        alias: String,
-        invalidatedByBiometricEnrollment: Boolean,
-    ): SecretKey = delegate.generateSecretKey(
-        alias = alias,
-        invalidatedByBiometricEnrollment = invalidatedByBiometricEnrollment,
-    )
-
-    /**
-     * Deletes key stored under [alias]. This is a no-op when key does not exist.
-     */
-    override fun deleteSecretKey(alias: String) = delegate.deleteSecretKey(alias = alias)
-
-    /**
-     * Checks whether fingerprint hardware is available.
-     */
-    override fun isDeviceSupportFingerprint(): Boolean = delegate.isDeviceSupportFingerprint()
-
-    /**
-     * Checks whether face-authentication hardware is available.
-     */
-    override fun isDeviceSupportFaceAuth(): Boolean = delegate.isDeviceSupportFaceAuth()
-
-    /**
-     * Checks whether at least one biometric method is supported.
-     */
-    override fun isDeviceSupportBiometric(): Boolean = delegate.isDeviceSupportBiometric()
-
-    /**
-     * Checks whether biometric credential is already enrolled.
-     */
-    override fun isBiometricEnrolled(): Boolean = delegate.isBiometricEnrolled()
-
-    /**
-     * Checks whether PIN/pattern/password is already enrolled.
-     */
-    override fun isDeviceCredentialEnrolled(): Boolean = delegate.isDeviceCredentialEnrolled()
-
-    /**
-     * Returns normalized status for [method].
-     */
-    override fun checkAuthenticatorStatus(method: MarkAuthenticatorMethod): MarkAuthenticationStatus =
-        delegate.checkAuthenticatorStatus(method = method)
-
-    /**
-     * Returns normalized status for secure authentication (strong biometric).
-     */
-    override fun checkSecureAuthentication(): MarkAuthenticationStatus = delegate.checkSecureAuthentication()
-
-    /**
-     * Returns `true` when [method] can authenticate in current device state.
-     */
-    override fun canAuthenticate(method: MarkAuthenticatorMethod): Boolean =
-        delegate.canAuthenticate(method = method)
-
-    /**
-     * Shows device credential authentication prompt.
-     */
-    override fun authenticateDeviceCredential(
-        activity: FragmentActivity,
-        title: String,
-        subTitle: String?,
-        description: String,
-        negativeText: String,
-        confirmationRequired: Boolean,
-        callback: WeakAuthenticationCallback,
-    ) = delegate.authenticateDeviceCredential(
-        activity = activity,
-        title = title,
-        subTitle = subTitle,
-        description = description,
-        negativeText = negativeText,
-        confirmationRequired = confirmationRequired,
-        callback = callback,
-    )
-
-    /**
-     * Shows weak biometric authentication prompt.
-     */
-    override fun authenticateBiometric(
-        activity: FragmentActivity,
-        title: String,
-        subTitle: String?,
-        description: String,
-        negativeText: String,
-        confirmationRequired: Boolean,
-        callback: WeakAuthenticationCallback,
-    ) = delegate.authenticateBiometric(
-        activity = activity,
-        title = title,
-        subTitle = subTitle,
-        description = description,
-        negativeText = negativeText,
-        confirmationRequired = confirmationRequired,
-        callback = callback,
-    )
-
-    /**
-     * Detects whether biometric enrollment changed after key creation for [alias].
-     */
-    @RequiresApi(Build.VERSION_CODES.M)
-    override fun isBiometricChanged(alias: String): Boolean = delegate.isBiometricChanged(alias = alias)
-
-    /**
-     * Authenticates using strong biometric and prepares cipher for encryption.
-     *
-     * This overload manages key and cipher internally.
-     */
-    @RequiresApi(Build.VERSION_CODES.M)
-    override fun secureAuthenticateBiometricEncrypt(
-        activity: FragmentActivity,
-        alias: String,
-        title: String,
-        subTitle: String?,
-        invalidatedByBiometricEnrollment: Boolean,
-        description: String,
-        negativeText: String,
-        confirmationRequired: Boolean,
-        callback: SecureAuthenticationEncryptCallback,
-    ) = delegate.secureAuthenticateBiometricEncrypt(
-        activity = activity,
-        alias = alias,
-        title = title,
-        subTitle = subTitle,
-        invalidatedByBiometricEnrollment = invalidatedByBiometricEnrollment,
-        description = description,
-        negativeText = negativeText,
-        confirmationRequired = confirmationRequired,
-        callback = callback,
-    )
-
-    /**
-     * Authenticates using strong biometric and prepares provided [cipher]/[secretKey] for encryption.
-     */
-    @RequiresApi(Build.VERSION_CODES.M)
-    override fun secureAuthenticateBiometricEncrypt(
-        activity: FragmentActivity,
-        title: String,
-        cipher: Cipher,
-        secretKey: SecretKey,
-        subTitle: String?,
-        description: String,
-        negativeText: String,
-        confirmationRequired: Boolean,
-        callback: SecureAuthenticationEncryptCallback,
-    ) = delegate.secureAuthenticateBiometricEncrypt(
-        activity = activity,
-        title = title,
-        cipher = cipher,
-        secretKey = secretKey,
-        subTitle = subTitle,
-        description = description,
-        negativeText = negativeText,
-        confirmationRequired = confirmationRequired,
-        callback = callback,
-    )
-
-    /**
-     * Authenticates using strong biometric and prepares cipher for decryption.
-     *
-     * This overload manages key and cipher internally.
-     */
-    @RequiresApi(Build.VERSION_CODES.M)
-    override fun secureAuthenticateBiometricDecrypt(
-        activity: FragmentActivity,
-        alias: String,
-        encodedIVKey: String,
-        title: String,
-        subTitle: String?,
-        description: String,
-        negativeText: String,
-        confirmationRequired: Boolean,
-        callback: SecureAuthenticationDecryptCallback,
-    ) = delegate.secureAuthenticateBiometricDecrypt(
-        activity = activity,
-        alias = alias,
-        encodedIVKey = encodedIVKey,
-        title = title,
-        subTitle = subTitle,
-        description = description,
-        negativeText = negativeText,
-        confirmationRequired = confirmationRequired,
-        callback = callback,
-    )
-
-    /**
-     * Authenticates using strong biometric and prepares provided [cipher]/[secretKey] for decryption.
-     */
-    @RequiresApi(Build.VERSION_CODES.M)
-    override fun secureAuthenticateBiometricDecrypt(
-        activity: FragmentActivity,
-        encodedIVKey: String,
-        cipher: Cipher,
-        secretKey: SecretKey,
-        title: String,
-        subTitle: String?,
-        description: String,
-        negativeText: String,
-        confirmationRequired: Boolean,
-        callback: SecureAuthenticationDecryptCallback,
-    ) = delegate.secureAuthenticateBiometricDecrypt(
-        activity = activity,
-        encodedIVKey = encodedIVKey,
-        cipher = cipher,
-        secretKey = secretKey,
-        title = title,
-        subTitle = subTitle,
-        description = description,
-        negativeText = negativeText,
-        confirmationRequired = confirmationRequired,
-        callback = callback,
-    )
-
-    /**
-     * Encrypts plaintext using authenticated [cipher], then returns Base64 encoded text.
-     */
-    override fun encrypt(cipher: Cipher, plainText: String): String =
-        delegate.encrypt(cipher = cipher, plainText = plainText)
-
-    /**
-     * Decrypts raw encrypted bytes using authenticated [cipher].
-     */
-    override fun decrypt(cipher: Cipher, encryptedText: ByteArray): String = delegate.decrypt(
-        cipher = cipher,
-        encryptedText = encryptedText,
-    )
-
-    /**
-     * Decrypts Base64 encoded text using authenticated [cipher].
-     */
-    override fun decrypt(cipher: Cipher, encryptedText: String): String = delegate.decrypt(
-        cipher = cipher,
-        encryptedText = encryptedText,
-    )
-}
