--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/domain/enums/MarkAuthenticationType.kt
@@ -1,7 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.domain.enums
-
-internal enum class MarkAuthenticationType {
-    BIOMETRIC_WEAK,
-    BIOMETRIC_STRONG,
-    DEVICE_CREDENTIAL,
-}
