--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/domain/enums/MarkAuthenticatorMethod.kt
@@ -1,16 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.domain.enums
-
-/**
- * Authentication methods supported by Mark Authenticator.
- */
-enum class MarkAuthenticatorMethod {
-    /**
-     * Weak biometric authentication (fingerprint, face, and equivalent weak classes).
-     */
-    BIOMETRIC,
-
-    /**
-     * Device credential authentication (PIN, pattern, or password).
-     */
-    DEVICE_CREDENTIAL,
-}
