--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/internal/dependency/capability/DeviceCapabilityDataSource.kt
@@ -1,10 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.internal.dependency.capability
-
-internal interface DeviceCapabilityDataSource {
-    val sdkInt: Int
-    fun hasSystemFeature(feature: String): Boolean
-    fun isFingerprintHardwareDetected(): Boolean
-    fun hasEnrolledFingerprints(): Boolean
-    fun isDeviceSecure(): Boolean
-    fun canAuthenticate(authenticator: Int): Int?
-}
