--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/internal/dependency/crypto/CipherDataSource.kt
@@ -1,7 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.internal.dependency.crypto
-
-import javax.crypto.Cipher
-
-internal interface CipherDataSource {
-    fun createCipher(): Cipher
-}
