--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/internal/dependency/crypto/DefaultCipherDataSource.kt
@@ -1,7 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.internal.dependency.crypto
-
-import javax.crypto.Cipher
-
-internal class DefaultCipherDataSource : CipherDataSource {
-    override fun createCipher(): Cipher = Cipher.getInstance("AES/GCM/NoPadding")
-}
