--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/internal/dependency/encoding/AndroidBase64DataSource.kt
@@ -1,8 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.internal.dependency.encoding
-
-import android.util.Base64
-
-internal class AndroidBase64DataSource : Base64DataSource {
-    override fun encode(raw: ByteArray): String = Base64.encodeToString(raw, Base64.NO_WRAP)
-    override fun decode(encoded: String): ByteArray = Base64.decode(encoded, Base64.NO_WRAP)
-}
