--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/internal/dependency/encoding/Base64DataSource.kt
@@ -1,6 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.internal.dependency.encoding
-
-internal interface Base64DataSource {
-    fun encode(raw: ByteArray): String
-    fun decode(encoded: String): ByteArray
-}
