--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/internal/dependency/prompt/PromptDataSource.kt
@@ -1,5 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.internal.dependency.prompt
-
-internal interface PromptDataSource {
-    fun authenticate(request: PromptRequest)
-}
