--- mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/internal/dependency/prompt/PromptRequest.kt
@@ -1,16 +1,0 @@
-package com.fadlurahmanfdev.mark_authenticator.internal.dependency.prompt
-
-import androidx.biometric.BiometricPrompt
-import androidx.fragment.app.FragmentActivity
-
-internal data class PromptRequest(
-    val activity: FragmentActivity,
-    val callback: BiometricPrompt.AuthenticationCallback,
-    val cryptoObject: BiometricPrompt.CryptoObject?,
-    val confirmationRequired: Boolean,
-    val authenticator: Int,
-    val title: String,
-    val subTitle: String?,
-    val description: String,
-    val negativeText: String,
-)
