+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/domain/enums/MarkAuthenticationType.kt
@@ -1,0 +1,7 @@
+package com.fadlurahmanfdev.mark_biometric.domain.enums
+
+internal enum class MarkAuthenticationType {
+    BIOMETRIC_WEAK,
+    BIOMETRIC_STRONG,
+    DEVICE_CREDENTIAL,
+}
