+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/domain/enums/MarkAuthenticatorMethod.kt
@@ -1,0 +1,16 @@
+package com.fadlurahmanfdev.mark_biometric.domain.enums
+
+/**
+ * Authentication methods supported by Mark Biometric.
+ */
+enum class MarkAuthenticatorMethod {
+    /**
+     * Weak biometric authentication (fingerprint, face, and equivalent weak classes).
+     */
+    BIOMETRIC,
+
+    /**
+     * Device credential authentication (PIN, pattern, or password).
+     */
+    DEVICE_CREDENTIAL,
+}
