+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/internal/dependency/crypto/AndroidSecretKeyDataSource.kt
@@ -1,0 +1,39 @@
+package com.fadlurahmanfdev.mark_biometric.internal.dependency.crypto
+
+import android.os.Build
+import android.security.keystore.KeyGenParameterSpec
+import android.security.keystore.KeyProperties
+import java.security.KeyStore
+import javax.crypto.KeyGenerator
+import javax.crypto.SecretKey
+
+internal class AndroidSecretKeyDataSource : SecretKeyDataSource {
+    override fun getSecretKey(alias: String): SecretKey? {
+        val keyStore = KeyStore.getInstance("AndroidKeyStore")
+        keyStore.load(null)
+        return keyStore.getKey(alias, null) as SecretKey?
+    }
+
+    override fun generateSecretKey(alias: String, invalidatedByBiometricEnrollment: Boolean): SecretKey {
+        val keyGenerator = KeyGenerator.getInstance(KeyProperties.KEY_ALGORITHM_AES, "AndroidKeyStore")
+        val spec = KeyGenParameterSpec.Builder(
+            alias,
+            KeyProperties.PURPOSE_ENCRYPT or KeyProperties.PURPOSE_DECRYPT,
+        ).apply {
+            setBlockModes(KeyProperties.BLOCK_MODE_GCM)
+            setEncryptionPaddings(KeyProperties.ENCRYPTION_PADDING_NONE)
+            setUserAuthenticationRequired(true)
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                setInvalidatedByBiometricEnrollment(invalidatedByBiometricEnrollment)
+            }
+        }.build()
+        keyGenerator.init(spec)
+        return keyGenerator.generateKey()
+    }
+
+    override fun deleteSecretKey(alias: String) {
+        val keyStore = KeyStore.getInstance("AndroidKeyStore")
+        keyStore.load(null)
+        keyStore.deleteEntry(alias)
+    }
+}
