+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/internal/dependency/crypto/CipherDataSource.kt
@@ -1,0 +1,7 @@
+package com.fadlurahmanfdev.mark_biometric.internal.dependency.crypto
+
+import javax.crypto.Cipher
+
+internal interface CipherDataSource {
+    fun createCipher(): Cipher
+}
