+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/internal/dependency/crypto/DefaultCipherDataSource.kt
@@ -1,0 +1,7 @@
+package com.fadlurahmanfdev.mark_biometric.internal.dependency.crypto
+
+import javax.crypto.Cipher
+
+internal class DefaultCipherDataSource : CipherDataSource {
+    override fun createCipher(): Cipher = Cipher.getInstance("AES/GCM/NoPadding")
+}
