+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/internal/dependency/crypto/SecretKeyDataSource.kt
@@ -1,0 +1,9 @@
+package com.fadlurahmanfdev.mark_biometric.internal.dependency.crypto
+
+import javax.crypto.SecretKey
+
+internal interface SecretKeyDataSource {
+    fun getSecretKey(alias: String): SecretKey?
+    fun generateSecretKey(alias: String, invalidatedByBiometricEnrollment: Boolean): SecretKey
+    fun deleteSecretKey(alias: String)
+}
