+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/internal/dependency/encoding/Base64DataSource.kt
@@ -1,0 +1,6 @@
+package com.fadlurahmanfdev.mark_biometric.internal.dependency.encoding
+
+internal interface Base64DataSource {
+    fun encode(raw: ByteArray): String
+    fun decode(encoded: String): ByteArray
+}
