+++ mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/internal/dependency/prompt/PromptDataSource.kt
@@ -1,0 +1,5 @@
+package com.fadlurahmanfdev.mark_biometric.internal.dependency.prompt
+
+internal interface PromptDataSource {
+    fun authenticate(request: PromptRequest)
+}
